+++ src/domain/interfaces/IFileImageRepository.ts
@@ -1,9 +1,9 @@
 import {ICrudRepository} from '@steroidsjs/nest/usecases/interfaces/ICrudRepository';
 import {FileImageModel} from '../models/FileImageModel';
-import FileStorageEnum from '../enums/FileStorageEnum';
+import {FileStorageNameType} from '../types/FileStorageNameType';
 
 export const IFileImageRepository = 'IFileImageRepository';
 
 export interface IFileImageRepository extends ICrudRepository<FileImageModel> {
-    getFilesPathsByStorageName: (storageName: FileStorageEnum) => Promise<string[] | null>;
+    getFilesPathsByStorageName: (storageName: FileStorageNameType) => Promise<string[] | null>,
 }