+++ src/domain/types/FileStorageNameType.ts
@@ -1,0 +1,3 @@
+import FileStorageEnum from '../enums/FileStorageEnum';
+
+export type FileStorageNameType = FileStorageEnum | string;